third_party/WebKit/LayoutTests/fast/css/getComputedStyle/getComputedStyle-webkit-columns-shorthand-expected.txt

Tests that the -webkit-columns shorthand is computed properly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS computedStyle.getPropertyValue('-webkit-columns') is "10px 2"
PASS computedStyle.getPropertyValue('-webkit-columns') is "10px auto"
PASS computedStyle.getPropertyValue('-webkit-columns') is "auto 2"
PASS computedStyle.getPropertyValue('-webkit-columns') is "auto 2"
PASS computedStyle.getPropertyValue('-webkit-columns') is "auto auto"
PASS computedStyle.getPropertyValue('-webkit-columns') is "10px auto"
PASS successfullyParsed is true

TEST COMPLETE
